+++ CHANGELOG.md
@@ -4,8 +4,12 @@ VC Verifier Changelog
 WIP
 ---
 
+
+1.5.3 (2023-06-14)
+---
+
 - always use id hash as the credential identifier
-- add proof challenge as query param
+- add proof challenge & domain as query param
 
 
 1.5.2 (2023-06-05)